Move logic out of KvssinkUtils due to global log4cplus initialization
